+++ src/py_bugger/cli/cli_utils.py
@@ -7,6 +7,9 @@ filenames.
 import os
 import sys
 from pathlib import Path
+import subprocess
+import shlex
+import shutil
 
 import click
 
@@ -29,8 +32,12 @@ def validate_config():
 
     if pb_config.target_file:
         _validate_target_file()
-        
+
+    # Update all options before running Git status checks. Info like target_dir
+    # is used for those checks.
     _update_options()
+
+    _validate_git_status()
 
 
 # --- Helper functions ___
@@ -47,6 +54,7 @@ def _update_options():
     # Make sure target_file is a Path.
     if pb_config.target_file:
         pb_config.target_file = Path(pb_config.target_file)
+
 
 def _validate_target_dir():
     """Make sure a valid directory was passed.
@@ -66,6 +74,7 @@ def _validate_target_dir():
         msg = cli_messages.msg_not_dir(path_target_dir)
         click.echo(msg)
         sys.exit()
+
 
 def _validate_target_file():
     """Make sure an appropriate file was passed.
@@ -87,5 +96,44 @@ def _validate_target_file():
         sys.exit()
     elif path_target_file.suffix != ".py":
         msg = cli_messages.msg_file_not_py(path_target_file)
+        click.echo(msg)
+        sys.exit()
+
+
+def _validate_git_status():
+    """Look for a clean Git status before introducing bugs."""
+    if pb_config.ignore_git_status:
+        return
+
+    _check_git_available()
+    _check_git_status()
+
+
+def _check_git_available():
+    """Quit with appropriate message if Git not available."""
+    if not shutil.which("git"):
+        click.echo(cli_messages.msg_git_not_available)
+        sys.exit()
+
+
+def _check_git_status():
+    """Make sure we're starting with a clean git status."""
+    if pb_config.target_file:
+        git_dir = pb_config.target_file.parent
+    else:
+        git_dir = pb_config.target_dir
+
+    cmd = "git status --porcelain"
+    cmd_parts = shlex.split(cmd)
+    output = subprocess.run(cmd_parts, cwd=git_dir, capture_output=True, text=True)
+
+    if "fatal: not a git repository" in output.stderr:
+        msg = cli_messages.msg_git_not_used(pb_config)
+        click.echo(msg)
+        sys.exit()
+
+    # `git status --porcelain` has no output when the status is clean.
+    if output.stdout or output.stderr:
+        msg = cli_messages.msg_unclean_git_status
         click.echo(msg)
         sys.exit()